# frontend/public/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentStat": {
        "description": "总览「按场景(Agent)分」的单个 Agent 聚合统计。\n\n「场景」= Agent(本项目铁律无场景类型枚举,用途由 Agent 决定)。通过率口径:\n分母 = 有评测结果的会话数(evaluated);分子 = passed 的会话数(优先人工复核 review_passed,\n回退 AI passed;dimension_score 模式无 passed → 按 overall_score>=阈值,由聚合侧折算)。\n未出评测结果(评测中/失败/未完成)的会话计入 total 但**不计入 evaluated 分母**,避免拉低通过率。",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name"
          },
          "completed": {
            "default": 0,
            "title": "Completed",
            "type": "integer"
          },
          "evaluated": {
            "default": 0,
            "title": "Evaluated",
            "type": "integer"
          },
          "pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Rate"
          },
          "passed": {
            "default": 0,
            "title": "Passed",
            "type": "integer"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "AgentStat",
        "type": "object"
      },
...
      "CandidateStatusOut": {
        "description": "候选人侧只见流程状态,不见任何评分/转写/录音(design contract 结果隔离)。",
        "properties": {
          "booked": {
            "default": false,
            "title": "Booked",
            "type": "boolean"
          },
          "engagement_id": {
            "title": "Engagement Id",
            "type": "string"
          },
          "meeting_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting End"
          },
          "meeting_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Start"
          },
          "slot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Id"
          },
          "stage": {
            "default": "not_booked",
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "engagement_id"
        ],
        "title": "CandidateStatusOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LlmCredentialStatusOut": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "status": {
            "enum": [
              "ok",
              "expiring",
              "expired",
              "not_configured",
              "not_applicable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "LlmCredentialStatusOut",
        "type": "object"
      },
...
      "SessionStatsOut": {
        "description": "总览统计聚合(spec:按场景分 + 通过率)。按 Agent 分组,按会话总数倒序。",
        "properties": {
          "agents": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AgentStat"
            },
            "title": "Agents",
            "type": "array"
          }
        },
        "title": "SessionStatsOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "VivaVoce control plane (real-time voice oral-exam). All /api/* endpoints require authentication.",
    "title": "VivaVoce Orchestrator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/candidate/status": {
      "get": {
        "description": "候选人侧状态(只见流程态,不见评分/转写/录音)。",
        "operationId": "candidate_status_api_candidate_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "in": "header",
            "name": "x-candidate-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Candidate-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Candidate Status",
        "tags": [
          "candidate"
        ]
      }
    },
...
    "/api/llm-credential-status": {
      "get": {
        "operationId": "get_llm_credential_status_api_llm_credential_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmCredentialStatusOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Llm Credential Status",
        "tags": [
          "llm-status"
        ]
      }
    },
    "/api/mcp": {
      "get": {
        "description": "MCP GET(可选 SSE 流):本实现不提供服务器推送。\n\ndesign contract:未认证 GET 返回 **401 + challenge**(而非裸 405)—— 若目标 client 先开 GET(SSE)探测再\n进发现流,405 可能使其不进 OAuth;返回带 `WWW-Authenticate` 的 401 让 OAuth-aware client 进发现。\n已认证但无 SSE 可推 → 405(transport 允许)。",
        "operationId": "mcp_get_api_mcp_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-delegation-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Delegation-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Get",
        "tags": [
          "mcp"
        ]
      },
      "post": {
        "description": "MCP JSON-RPC 端点。Bearer / 委托 token 鉴权 → 解析 staff → 处理 initialize/tools.list/tools.call。",
        "operationId": "mcp_endpoint_api_mcp_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-delegation-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Delegation-Token"
            }
          },
          {
            "in": "header",
            "name": "mcp-protocol-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mcp-Protocol-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Endpoint",
        "tags": [
          "mcp"
        ]
      }
    },
...
    "/api/sessions/stats": {
      "get": {
        "description": "总览「按场景(Agent)分 + 通过率」聚合。staff 只统计自己的会话(归属隔离,不信前端)。\n\n通过率 = passed 会话 / 有评测结果的会话(evaluated 为分母);未出结果的会话计入 total 不计 evaluated。\nSession×Result 按 session_id join;结果归属由「可见 session 集合」保证(不单独信任 Results scan)。",
        "operationId": "session_stats_api_sessions_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-delegation-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Delegation-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Session Stats",
        "tags": [
          "sessions"
        ]
      }
    },
...
  }
}